Switch allocation benchmarks to GHA
Motivation:

This will allow us to decommission the old CI hooks.

Modifications:

* Migrate the existing tests to use the package-benchmark framework.
* Delete the obsolete integration tests and docker files.

Result:

Allocation counts are checked using GitHub Actions.
