+++ .github/workflows/pull_request.yml
@@ -20,6 +20,12 @@ jobs:
             linux_nightly_6_0_arguments_override: "--explicit-target-dependency-import-check error -Xswiftc -require-explicit-sendable"
             linux_nightly_main_arguments_override: "--explicit-target-dependency-import-check error -Xswiftc -require-explicit-sendable"
 
+    benchmarks:
+        name: Benchmarks
+        uses: apple/swift-nio/.github/workflows/benchmarks.yml@main
+        with:
+            benchmark_package_path: "Benchmarks"
+
     cxx-interop:
         name: Cxx interop
         uses: apple/swift-nio/.github/workflows/cxx_interop.yml@main
