+++ .github/workflows/scheduled.yml
@@ -15,3 +15,9 @@ jobs:
             linux_6_0_arguments_override: "--explicit-target-dependency-import-check error -Xswiftc -require-explicit-sendable"
             linux_nightly_6_0_arguments_override: "--explicit-target-dependency-import-check error -Xswiftc -require-explicit-sendable"
             linux_nightly_main_arguments_override: "--explicit-target-dependency-import-check error -Xswiftc -require-explicit-sendable"
+
+    benchmarks:
+        name: Benchmarks
+        uses: apple/swift-nio/.github/workflows/benchmarks.yml@main
+        with:
+            benchmark_package_path: "Benchmarks"
