+++ Benchmarks/Thresholds/nightly-6.0/NIOSSHBenchmarks.ManySmallCommandsPerConnection.p90.json
@@ -1,0 +1,3 @@
+{
+  "mallocCountTotal" : 192621
+}
