+++ Benchmarks/Thresholds/nightly-6.0/NIOSSHBenchmarks.OneCommandPerConnection.p90.json
@@ -1,0 +1,3 @@
+{
+  "mallocCountTotal" : 811999
+}
