+++ Benchmarks/Thresholds/nightly-6.0/NIOSSHBenchmarks.StreamingLargeMessageInSmallChunks.p90.json
@@ -1,0 +1,3 @@
+{
+  "mallocCountTotal" : 40790
+}
