+++ Benchmarks/Thresholds/nightly-main/NIOSSHBenchmarks.ManySmallCommandsPerConnection.p90.json
@@ -1,0 +1,3 @@
+{
+  "mallocCountTotal" : 192621
+}
