+++ Benchmarks/Thresholds/nightly-main/NIOSSHBenchmarks.StreamingLargeMessageInSmallChunks.p90.json
@@ -1,0 +1,3 @@
+{
+  "mallocCountTotal" : 40790
+}
