--- docker/docker-compose.2204.59.yaml
@@ -1,21 +1,0 @@
-version: "3"
-
-services:
-
-  runtime-setup:
-    image: swift-nio-ssh:22.04-5.9
-    build:
-      args:
-        ubuntu_version: "jammy"
-        swift_version: "5.9"
-
-  test:
-    image: swift-nio-ssh:22.04-5.9
-    environment:
-      - MAX_ALLOCS_ALLOWED_client_server_many_small_commands_per_connection=198700
-      - MAX_ALLOCS_ALLOWED_client_server_one_command_per_connection=862050
-      - MAX_ALLOCS_ALLOWED_client_server_streaming_large_message_in_small_chunks=42850
-      - IMPORT_CHECK_ARG=--explicit-target-dependency-import-check error
-      - WARN_AS_ERROR_ARG=-Xswiftc -warnings-as-errors
-  shell:
-    image: swift-nio-ssh:22.04-5.9
