--- docker/docker-compose.2204.main.yaml
@@ -1,21 +1,0 @@
-version: "3"
-
-services:
-
-  runtime-setup:
-    image: swift-nio-ssh:22.04-main
-    build:
-      args:
-        base_image: "swiftlang/swift:nightly-main-jammy"
-
-  test:
-    image: swift-nio-ssh:22.04-main
-    environment:
-      - MAX_ALLOCS_ALLOWED_client_server_many_small_commands_per_connection=192700
-      - MAX_ALLOCS_ALLOWED_client_server_one_command_per_connection=852050
-      - MAX_ALLOCS_ALLOWED_client_server_streaming_large_message_in_small_chunks=40850
-      - IMPORT_CHECK_ARG=--explicit-target-dependency-import-check error
-      - WARN_AS_ERROR_ARG=-Xswiftc -warnings-as-errors
-
-  shell:
-    image: swift-nio-ssh:22.04-main
