--- docker/docker-compose.yaml
@@ -1,33 +1,0 @@
-# this file is not designed to be run directly
-# instead, use the docker-compose.<os>.<swift> files
-# eg docker-compose -f docker/docker-compose.yaml -f docker/docker-compose.1804.51.yaml run test
-version: "3"
-
-services:
-
-  runtime-setup:
-    image: swift-nio-ssh:default
-    build:
-      context: .
-      dockerfile: Dockerfile
-
-  common: &common
-    image: swift-nio-ssh:default
-    depends_on: [runtime-setup]
-    volumes:
-      - ~/.ssh:/root/.ssh
-      - ..:/code:z
-    working_dir: /code
-    cap_drop:
-      - CAP_NET_RAW
-      - CAP_NET_BIND_SERVICE
-
-  test:
-    <<: *common
-    command: /bin/bash -xcl "./scripts/integration_tests.sh $${INTEGRATION_TESTS_ARG-}"
-
-  # util
-
-  shell:
-    <<: *common
-    entrypoint: /bin/bash
